Rename curves to Pallas/Vesta (Pasta).
