+++ benches/arithmetic.rs
@@ -5,7 +5,7 @@ extern crate halo2;
 use crate::arithmetic::{small_multiexp, FieldExt};
 use crate::poly::commitment::Params;
 use crate::transcript::DummyHash;
-use crate::tweedle::{EqAffine, Fp, Fq};
+use crate::pasta::{EqAffine, Fp, Fq};
 use halo2::*;
 
 use criterion::{black_box, Criterion};
