+++ src/pasta.rs
@@ -1,0 +1,13 @@
+//! This module contains implementations for the Pallas and Vesta elliptic curve
+//! groups.
+
+#[macro_use]
+mod macros;
+mod curves;
+mod fields;
+
+pub mod pallas;
+pub mod vesta;
+
+pub use curves::*;
+pub use fields::*;
