--- src/tweedle.rs
@@ -1,13 +1,0 @@
-//! This module contains implementations for the Tweedledum and Tweedledee
-//! elliptic curve groups.
-
-#[macro_use]
-mod macros;
-mod curves;
-mod fields;
-
-pub mod dee;
-pub mod dum;
-
-pub use curves::*;
-pub use fields::*;
